Consolidate tracking code into single file for DigitalOcean Functions
Problem: DigitalOcean Functions was treating every .py file as a separate
function, causing deployment errors:
- tracking.py was deployed as an unwanted function
- requirements.txt failed with "runtime type could not be determined"

Solution: Combined tracking.py into __main__.py as a single file
- All tracking functions now in one file (no import issues)
- Only one function deploys (latest-installer/__main__)
- Graceful fallback if facebook-business SDK not available
- requirements.txt properly used for dependencies

This is the standard pattern for serverless functions - single file
deployment avoids module import complexities.

Tested locally - imports work correctly.
